+++ src/main/java/model/result/Result.java
@@ -1,4 +1,4 @@
-package model;
+package model.result;
 
 public record Result(String name) {
 }
